+++ acceptance/openstack/client_test.go
@@ -8,7 +8,13 @@ import (
 	"time"
 
 	"github.com/gophercloud/gophercloud"
+	"github.com/gophercloud/gophercloud/acceptance/clients"
+	"github.com/gophercloud/gophercloud/acceptance/tools"
 	"github.com/gophercloud/gophercloud/openstack"
+	"github.com/gophercloud/gophercloud/openstack/identity/v3/credentials"
+	"github.com/gophercloud/gophercloud/openstack/identity/v3/extensions/ec2tokens"
+	"github.com/gophercloud/gophercloud/openstack/identity/v3/tokens"
+	th "github.com/gophercloud/gophercloud/testhelper"
 )
 
 func TestAuthenticatedClient(t *testing.T) {
@@ -38,6 +44,68 @@ func TestAuthenticatedClient(t *testing.T) {
 	} else {
 		t.Logf("Located a storage service at endpoint: [%s]", storage.Endpoint)
 	}
+}
+
+func TestEC2AuthMethod(t *testing.T) {
+	client, err := clients.NewIdentityV3Client()
+	th.AssertNoErr(t, err)
+
+	ao, err := openstack.AuthOptionsFromEnv()
+	th.AssertNoErr(t, err)
+
+	authOptions := tokens.AuthOptions{
+		Username:   ao.Username,
+		Password:   ao.Password,
+		DomainName: ao.DomainName,
+		DomainID:   ao.DomainID,
+		// We need a scope to get the token roles list
+		Scope: tokens.Scope{
+			ProjectID:   ao.TenantID,
+			ProjectName: ao.TenantName,
+			DomainID:    ao.DomainID,
+			DomainName:  ao.DomainName,
+		},
+	}
+	token, err := tokens.Create(client, &authOptions).Extract()
+	th.AssertNoErr(t, err)
+	tools.PrintResource(t, token)
+
+	user, err := tokens.Get(client, token.ID).ExtractUser()
+	th.AssertNoErr(t, err)
+	tools.PrintResource(t, user)
+
+	project, err := tokens.Get(client, token.ID).ExtractProject()
+	th.AssertNoErr(t, err)
+	tools.PrintResource(t, project)
+
+	createOpts := credentials.CreateOpts{
+		ProjectID: project.ID,
+		Type:      "ec2",
+		UserID:    user.ID,
+		Blob:      "{\"access\":\"181920\",\"secret\":\"secretKey\"}",
+	}
+
+	// Create a credential
+	credential, err := credentials.Create(client, createOpts).Extract()
+	th.AssertNoErr(t, err)
+
+	// Delete a credential
+	defer credentials.Delete(client, credential.ID)
+	tools.PrintResource(t, credential)
+
+	newClient, err := clients.NewIdentityV3UnauthenticatedClient()
+	th.AssertNoErr(t, err)
+
+	var ec2AuthOptions tokens.AuthOptionsBuilder
+	ec2AuthOptions = &ec2tokens.AuthOptions{
+		Access: "181920",
+		Secret: "secretKey",
+	}
+
+	err = openstack.AuthenticateV3(newClient.ProviderClient, ec2AuthOptions, gophercloud.EndpointOpts{})
+	th.AssertNoErr(t, err)
+
+	tools.PrintResource(t, newClient.TokenID)
 }
 
 func TestReauth(t *testing.T) {